requirements.txt
requests
pychef